引き戸の開閉と列挙体による制御
