+++ Assets/Plugins/Horror_House/Prefabs/Furniture/Kitchen_Module_01.prefab
@@ -343,7 +343,7 @@ GameObject:
   - component: {fileID: 1668865635367505327}
   - component: {fileID: 4618194251627315326}
   - component: {fileID: 3840020439163885840}
-  - component: {fileID: 2809680417905769901}
+  - component: {fileID: 939277082555368378}
   m_Layer: 10
   m_Name: Kitchen_Module_01_Door_LOD0
   m_TagString: Untagged
@@ -438,7 +438,7 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 0.057510257, y: 0.71482223, z: 0.5398478}
   m_Center: {x: 0.028754652, y: 0.35741112, z: -0.26748154}
---- !u!114 &2809680417905769901
+--- !u!114 &939277082555368378
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -450,4 +450,5 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d8c42b0ddf104834eba065e2435c2642, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  drawType: 0
   _initialStateOpen: 0
