+++ Assets/Plugins/Horror_House/Prefabs/Furniture/Kitchen_Module_02.prefab
@@ -202,6 +202,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d8c42b0ddf104834eba065e2435c2642, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  drawType: 0
   _initialStateOpen: 0
 --- !u!1 &2408203480924819138
 GameObject:
@@ -549,7 +550,7 @@ GameObject:
   - component: {fileID: 549536283278056646}
   - component: {fileID: 1998219190220666786}
   - component: {fileID: 151806117542822374}
-  - component: {fileID: 4633420085695420702}
+  - component: {fileID: 5311988785827351437}
   m_Layer: 10
   m_Name: Kitchen_Module_02,1_Door_LOD0
   m_TagString: Untagged
@@ -644,7 +645,7 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 0.057510376, y: 0.71482223, z: 0.5433772}
   m_Center: {x: 0.028755188, y: 0.35741112, z: 0.2716886}
---- !u!114 &4633420085695420702
+--- !u!114 &5311988785827351437
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -653,7 +654,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 9221164403772059779}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9ec193949eecbb74292634befa4f9a15, type: 3}
+  m_Script: {fileID: 11500000, guid: d8c42b0ddf104834eba065e2435c2642, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  drawType: 1
   _initialStateOpen: 0
