+++ Assets/Plugins/Horror_House/Prefabs/Furniture/Kitchen_Module_04.prefab
@@ -12,7 +12,8 @@ GameObject:
   - component: {fileID: 3205675935714695594}
   - component: {fileID: 5742049653446162148}
   - component: {fileID: 3496554124699360338}
-  m_Layer: 0
+  - component: {fileID: 423594033411844565}
+  m_Layer: 10
   m_Name: Kitchen_Module_04_Box_LOD0
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -106,6 +107,20 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 0.42486763, y: 0.17768592, z: 0.54242754}
   m_Center: {x: 0.21243382, y: 0.08150622, z: -0.26629758}
+--- !u!114 &423594033411844565
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 179157519393018664}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d8c42b0ddf104834eba065e2435c2642, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  drawType: 2
+  _initialStateOpen: 0
 --- !u!1 &1156674040278715314
 GameObject:
   m_ObjectHideFlags: 0
@@ -642,4 +657,5 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d8c42b0ddf104834eba065e2435c2642, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  drawType: 0
   _initialStateOpen: 0
