--- Assets/Scripts/Scenes/InGame/Stage/StageRack_Left.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9ec193949eecbb74292634befa4f9a15
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
